requirements.txt
numpy
matplotlib
scipy
netCDF4
pooch
Cython
setuptools
cartopy
cmweather
xarray
mda-xdrlib
fsspec
s3fs
pandas
geopandas
pint
gdal==3.8.4
scikit-learn
scikit-image
scikit-fuzzy
tensorflow==2.16.2
tensorflow-probability==0.24.0
#tensorflow==2.20.0
#tensorflow-probability==0.25.0
tf-keras
cfgrib
xradar
wradlib
arm_pyart
metpy
pydda
rpy2
getgfs
